For: "DEV-12441 Save and Close does not save changed image."
When `this.closePage();` is called, it must deconstruct the view
objects, and it cancels any `workletSaved` callbacks.  So, we have to
move the `worklet.saveFiles();` call into the actual 'save' machinery.

This also fixes another bug where, if you upload a changed image,
and then switch tabs, the image would be broken.  This reloads the
image.

#backbone #js_helpers #jasmine_tests