A final change for: "DEV-11637 The "Help and Support" dialog should have a button that lets users download log files" ...
Guohui Gao found a bug: "when i download logs for registered_hdfs_data_sources is empty"

The reason it's empty is that the app was trying to get the data by downloading it via curl from the json API.  The
problem is, there's an authentication system around the API and you can only use it if you are a logged in user.

So, instead of using the API, I am just getting the data directly out of the database via the HdfsDataSource model,
formatting it into JSON, and putting it into a file.  This will probably not be as verbose as the API would have been,
but it's simple and it's a lot better than nothing.  I am sure we will be refining this further, shortly.